WIP: Update Rackspace so node talks to apiserver
Kubernetes project has decided that it is better if kubelet
and kube-proxy use the apiserver REST interface to get and
set resources instead of accessing resource keys in etcd directly.
This is necessary to support kubelet reporting of events,
and also encapsulates the apiserver store details.

This means that the kubelet and kube-proxy need to know the
apiserver host(s) via a flag.

Since the Rackspace config already used etcd to advertise the
minions to the controller-manager, I used the same pattern to advertise
the apiserver(s) to the minions.

Setting --public_address_override=$private_ipv4 is intended to ensure that
the master serves its http interface on the right ethernet device, since I think
there are two on a droplet.

The new apiserver-advertiser.service puts the IPs of any apiservers in etcd.

The kubelet and kube-proxy now take an environment file which contains
the list of apiserver IPs, and that env var goes into a flag.  The
etcd_servers argument is removed -- the point is for these binaries
to not access etcd.

The new apiserver-finder.service watches for changes in etcd and
restarts kubelet and kube proxy when there are new apiservers.