use $public_ipv4:6443 for apiservers
Accessing the apiserver via read-only or https
settings must be done via the public interface since
that's where they listen and I don't see a way to move
them to another elsewhere.

The GCE implementation uses the https(port 6443) for
this communication. We'll follow the same pattern but
still need to implement a certificate to be served. Using
the read-only port resulted in "failure to write" errors
which makes sense.

The python fixes were to force python2.7 since those
calls don't work with python3+.